examples/teaching_colony/requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
jsonschema>=4.0
pytest>=7.0
